pyproject.toml
[project]
name = "ditto-client"
version = "0.2.0"
description = "Eclipse Ditto Python Client"
readme = "README.md"
authors = [
    { name = "Sachdeva, Kapil"}
]
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "microsoft-kiota-abstractions>=1.9.7",
    "microsoft-kiota-http>=1.9.7",
    "microsoft-kiota-serialization-form>=1.9.7",
    "microsoft-kiota-serialization-json>=1.9.7",
    "microsoft-kiota-serialization-multipart>=1.9.7",
    "microsoft-kiota-serialization-text>=1.9.7",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "typer>=0.16.0",
    "rich>=14.0.0",
    "python-dotenv>=1.1.1",
]

classifiers = [
  "Intended Audience :: Developers",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: Apache Software License",
]

[project.urls]
repository = "https://github.com/ksachdeva/ditto-client"

[project.scripts]
ditto-client = "ditto_client.__main__:cli_app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-xdist>=3.6.1",
    "pytest-mock>=3.14.0",
    "poethepoet>=0.30.0",
    "packaging>=24.2",
    "pyright>=1.1.378",
    "mypy>=1.10.0",
    "ruff>=0.4.8",
    "pre-commit>=4.3.0",
    "rich>=14.1.0",
]

[tool.ruff]
line-length = 120
fix = true
target-version = "py311"

exclude = [
    "src/ditto_client/api",
    "src/ditto_client/devops",
    "src/ditto_client/models",
    "src/ditto_client/ditto_client.py",

]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["E", "F", "W", "B", "Q", "I", "ASYNC", "T20"]
ignore = ["F401", "E501"]

[tool.ruff.lint.flake8-tidy-imports]
[tool.ruff.lint.flake8-tidy-imports.banned-api]
"unittest".msg = "Use `pytest` instead."

[tool.ruff.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.params.Depends", "fastapi.Query", "fastapi.params.Query"]

[tool.mypy]
strict = true
python_version = "3.11"
ignore_missing_imports = true

# from https://blog.wolt.com/engineering/2021/09/30/professional-grade-mypy-configuration/
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
show_error_codes = true
warn_unused_ignores = false

disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_any_unimported = false

[tool.pyright]
include = ["src", "tests", "samples"]
# typeCheckingMode = "strict"
reportUnnecessaryIsInstance = false
reportMissingTypeStubs = false

[tool.poe.tasks.generate-kiota-client]
help = "Generate the client code using Kiota"
cmd = "docker run -v $PWD/src/ditto_client/generated:/app/output -v $PWD/assets/ditto/openapi/ditto-api-2.yml:/app/openapi.yaml mcr.microsoft.com/openapi/kiota generate -l python -c DittoClient"
